Added Class to Store Values
Now that the file input has been obtained and properly delimited, we can store the values in a class for ease of access and further functionality.
Using multiple TryParse statements, the strings were converted to the proper format
I learned about the differences between the functions of the "ref" and "out" keywords and how they work on reference data (such as objects and strings) and aren't required for structs and basic data types like integers.